1. Add network connection related to User 1.1 sign up 1.2 login 1.3 checking duplicate 1.4 save new user's info 2. Add network connection related to Map 2.1 There is an error so we need to check this part again.
