+++ app/src/main/java/com/kop/daegudot/Network/User/UserResponse.java
@@ -1,0 +1,11 @@
+package com.kop.daegudot.network;
+
+public class UserResponse {
+    public Long id;
+    public String email;
+    public String nickname;
+    public String password;
+    public String token;
+    public char type;
+
+}
