--- app/src/main/java/com/kop/daegudot/network/UserResponse.java
@@ -1,7 +1,0 @@
-package com.kop.daegudot.network;
-
-public class UserResponse {
-    private String email;
-    private String nickname;
-    private char type;
-}
